+++ client/src/components/Login/Login.jsx
@@ -29,7 +29,7 @@ const Login = ({ setIsAuth }) => {
       cookies.set("auth-token", result.user.refreshToken);
       setIsAuth(true);
       console.log(auth);
-      axios.put('http://localhost:3001/geton',{ uid, on:true} )
+      axios.put('http://localhost:3001/geton',{ uid, is: "on"} )
     } catch (error) {
       throw Error(error)
     }
@@ -43,7 +43,7 @@ const Login = ({ setIsAuth }) => {
       const result = await signInWithEmailAndPassword(auth, email, password);
       cookies.set("auth-token", result.user.refreshToken);
       setIsAuth(true);
-      axios.put('http://localhost:3001/geton',{ uid, on:true} )
+      axios.put('http://localhost:3001/geton',{ uid, is: "on"} )
     } catch (error) {
       throw Error(error)
     }
